+++ src/main/java/roomescape/dto/ReservationCreateDto.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.dto;
 
 public class ReservationCreateDto {
     private final String name;
